+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add more usage examples to the `about` documentation page.
 - Added `npm run start:nolint` script to disable linting for quickly prototyping code.
 - Added new file type and data type constants.
+- Added loaders for `obsFeatureMatrix.clusters.json`
 - Added loaders for `obsFeatureMatrix.genes.json`
 - Added loaders for `obsLabels.cells.json`, `obsEmbedding.cells.json`, `obsLocations.cells.json`, `obsSegmentations.cells.json`